deactivate `autoserve` by default
The automatic serving of uploaded images/files was very surpsing, as it
was badly documented.

It could even lead to data leak, as it was also activated by default.

Now, it is an opt-in, so if you want the behavior of `Flask-Uploads`,
you need to set `UPLOADS_AUTOSERVE=True`

This is the first major backwards incompatibility with `Flask-Uploads`.